[WalletModel] Stop polling balance if shutdown was requested.
Essentially the balance timer was still alive when the app's shutdown was requested. Causing a segfault as soon as it tries to get the best block hash from the already destroyed clientModel object.

This commit introduces an internal shutdownRequested check inside the polling method and for the sake of completeness, the timer stop when the WalletModel object gets destroyed.