gui: notify the user early if any of the recipients is invalid.
As we are already validating the entire recipient data before adding it to the temporal recipients vector (address and amount), the 'SendCoinsEntry' validation function can just return the proper error code, so it can be processed and the user can be notified right away.

Without requiring to unlock the wallet, nor walk through the entire recipients vector later again on the flow just to validate the same things twice.