wallet: do not check for external input solvabilty with wallet's data
If we have an external input that the wallet knows about, we will be able
to get the parent transaction but not be able to solve it.

So, instead of ask if the parent tx exist in the wallet to know whether
we own the output or not. We can use `IsMine(outpoint)` directly.